Added 'Bounds', Collisions, and Lose Conditions
Each entity now has a 'Bounds' that it can go outside of, but will no
longer be rendered if outside.

You can now find whether any two entities are colliding using
Entity.IsCollidingWith().

You now lose the game if the player collides with one of the enemies.

On loss, a Game Over! screen is displayed showing your total score for
that game.