Balance
Changed the number of enemies in-game from 20 to 15 to make the game
easier in the later stages.

Changed the max speed of enemies from 125 to a random value between 75
and 100.

This change should have been in the last commit, but GitHub didn't like
the idea of that, so it didn't happen.